improv: move about to a widget
