+++ examples/about/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 tracing = "0.1.37"
 tracing-subscriber = "0.3.17"
 tracing-log = "0.2.0"
+open = "5.3.0"
 
 [dependencies.libcosmic]
 path = "../../"
